test(domain): split FSM and transition helper tests
